Runtime: O(n^3). # Cols * # Rows * max(# Rows, # Cols)
The approach here is to keep the diagonal checking idea the same, but starting it at different points. The diagonal checking method applied here to is to go up and left until it is not possible anymore, or a value that does not match the target is found. The diagonal check is applied to the last row, as the only direction to check is upward, and to fulfill all possible diagonals, the values in the rightmost column are checked as targets, since there are no values to the right that need to be checked. From here, all diagonals will be found.