Runtime: O(n). Worst Case there is carry over to all values.
The approach here is to look at all possible cases that can occur when adding 1. The code is separated by cases. The overarching case is to check if all the values in digits are 9's, because the addition here is unique in that it increases the length of the list. If digits does not fit that case, then it is checked if there is an addition without carrying the 1 (adding 1 to not 9). Finally is the case for addition with carrying the 1, and the code is implemented to account for that.