Runtime: O(n). For loop goes through characters in the string
The first part of the code sets up acceptable characters that can be kept in strings. All white spaces and special characters are then removed, and uppercase letters are converted to lowercase. A palindrome is checked by comparing the first and last characters in a string, then moving inwards towards the middle. If a mismatch is found, return false.